extend SegmentRelocator instead TableRebalancer
