+++ pinot-common/src/main/java/org/apache/pinot/common/messages/SegmentReloadMessage.java
@@ -21,7 +21,6 @@ package org.apache.pinot.common.messages;
 import com.google.common.base.Preconditions;
 import java.util.List;
 import java.util.UUID;
-import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import org.apache.helix.model.Message;
 import org.apache.helix.zookeeper.datamodel.ZNRecord;
@@ -35,10 +34,9 @@ public class SegmentReloadMessage extends Message {
   public static final String RELOAD_SEGMENT_MSG_SUB_TYPE = "RELOAD_SEGMENT";
 
   private static final String FORCE_DOWNLOAD_KEY = "forceDownload";
-  private static final String SEGMENT_LIST_KEY = "segmentList";
+  private static final String SEGMENT_NAMES = "segmentNames";
 
-  public SegmentReloadMessage(@Nonnull String tableNameWithType, @Nullable List<String> segmentNames,
-      boolean forceDownload) {
+  public SegmentReloadMessage(String tableNameWithType, List<String> segmentNames, boolean forceDownload) {
     super(MessageType.USER_DEFINE_MSG, UUID.randomUUID().toString());
     setResourceName(tableNameWithType);
     setMsgSubType(RELOAD_SEGMENT_MSG_SUB_TYPE);
@@ -47,12 +45,10 @@ public class SegmentReloadMessage extends Message {
 
     ZNRecord znRecord = getRecord();
     znRecord.setBooleanField(FORCE_DOWNLOAD_KEY, forceDownload);
-    if (segmentNames != null) {
-      if (segmentNames.size() == 1) {
-        setPartitionName(segmentNames.get(0));
-      } else {
-        znRecord.setListField(SEGMENT_LIST_KEY, segmentNames);
-      }
+    znRecord.setListField(SEGMENT_NAMES, segmentNames);
+    // TODO: use the new List field and deprecate the partition name in next release.
+    if (!segmentNames.isEmpty()) {
+      setPartitionName(segmentNames.get(0));
     }
   }
 
@@ -67,7 +63,8 @@ public class SegmentReloadMessage extends Message {
     return getRecord().getBooleanField(FORCE_DOWNLOAD_KEY, false);
   }
 
+  @Nullable
   public List<String> getSegmentList() {
-    return getRecord().getListField(SEGMENT_LIST_KEY);
+    return getRecord().getListField(SEGMENT_NAMES);
   }
 }
